refactor: Move tax calculation to payment-entry-page
