+++ src/ui/purchases-ui.svelte
@@ -12,7 +12,6 @@
     PurchaseFlowError,
     PurchaseFlowErrorCode,
     PurchaseOperationHelper,
-    TaxCalculationError,
   } from "../helpers/purchase-operation-helper";
 
   import { type RedemptionInfo } from "../entities/redemption-info";
@@ -21,13 +20,7 @@
     Translator,
   } from "./localization/translator";
   import { translatorContextKey } from "./localization/constants";
-  import {
-    type PriceBreakdown,
-    type TaxCustomerDetails,
-    type CurrentPage,
-    TaxCalculationStatus,
-    TaxCalculationPendingReason,
-  } from "./ui-types";
+  import { type CurrentPage } from "./ui-types";
   import PurchasesUiInner from "./purchases-ui-inner.svelte";
   import { type ContinueHandlerParams } from "./ui-types";
   import { type IEventsTracker } from "../behavioural-events/events-tracker";
@@ -36,8 +29,6 @@
   import type { PurchaseMetadata } from "../entities/offerings";
   import { writable } from "svelte/store";
   import { type GatewayParams } from "../networking/responses/stripe-elements";
-  import { ALLOW_TAX_CALCULATION_FF } from "../helpers/constants";
-  import type { TaxBreakdown } from "../networking/responses/checkout-calculate-tax-response";
   import { validateEmail } from "../helpers/validators";
 
   interface Props {
@@ -88,31 +79,10 @@
   let lastError: PurchaseFlowError | null = $state(null);
   const productId = rcPackage.webBillingProduct.identifier ?? null;
 
-  let currentPage: CurrentPage | null = $state(null);
+  let currentPage: CurrentPage = $state("payment-entry-loading");
   let redemptionInfo: RedemptionInfo | null = $state(null);
   let operationSessionId: string | null = $state(null);
   let gatewayParams: GatewayParams = $state({});
-
-  let taxCalculationStatus: TaxCalculationStatus = $state("disabled");
-  let pendingReason: TaxCalculationPendingReason | null = $state(null);
-  let taxAmountInMicros: number | null = $state(null);
-  let taxBreakdown: TaxBreakdown[] | null = $state(null);
-  let totalExcludingTaxInMicros: number | null = $state(
-    productDetails.currentPrice.amountMicros,
-  );
-  let totalAmountInMicros: number | null = $state(
-    productDetails.currentPrice.amountMicros,
-  );
-
-  let priceBreakdown: PriceBreakdown = $derived({
-    currency: productDetails.currentPrice.currency,
-    totalAmountInMicros,
-    totalExcludingTaxInMicros,
-    taxCalculationStatus,
-    pendingReason,
-    taxAmountInMicros,
-    taxBreakdown,
-  });
 
   // Setting the context for the Localized components
   let translator: Translator = new Translator(
@@ -152,8 +122,6 @@
       return;
     }
 
-    currentPage = "payment-entry-loading";
-
     purchaseOperationHelper
       .checkoutStart(
         appUserId,
@@ -182,19 +150,6 @@
         lastError = null;
         currentPage = "payment-entry";
         gatewayParams = result.gateway_params;
-      })
-      .then(async (result) => {
-        // If tax collection is enabled, we start in the state `disabled`
-        if (
-          ALLOW_TAX_CALCULATION_FF &&
-          brandingInfo?.gateway_tax_collection_enabled
-        ) {
-          await refreshTaxCalculation();
-        }
-        return result;
-      })
-      .catch((e: PurchaseFlowError) => {
-        handleError(e);
       });
   });
 
@@ -227,59 +182,6 @@
     currentPage = "success";
   };
 
-  async function refreshTaxCalculation(
-    taxCustomerDetails: TaxCustomerDetails | undefined = undefined,
-  ) {
-    if (taxCalculationStatus !== "disabled") {
-      taxCalculationStatus = "loading";
-    }
-
-    const taxCalculation = await purchaseOperationHelper.checkoutCalculateTax(
-      taxCustomerDetails?.countryCode,
-      taxCustomerDetails?.postalCode,
-    );
-
-    if (taxCalculation.error) {
-      switch (taxCalculation.error) {
-        case TaxCalculationError.Pending:
-          taxCalculationStatus = "pending";
-          pendingReason = null;
-          break;
-        case TaxCalculationError.Disabled:
-          taxCalculationStatus = "disabled";
-          pendingReason = null;
-          break;
-        case TaxCalculationError.InvalidLocation:
-          taxCalculationStatus = "pending";
-          pendingReason = "invalid_postal_code";
-          break;
-        default:
-          handleError(
-            new PurchaseFlowError(
-              PurchaseFlowErrorCode.ErrorSettingUpPurchase,
-              "Unknown error without state set.",
-            ),
-          );
-      }
-
-      return;
-    }
-
-    const { data } = taxCalculation;
-
-    taxCalculationStatus = "calculated";
-    taxAmountInMicros = data.tax_amount_in_micros;
-    totalExcludingTaxInMicros = data.total_excluding_tax_in_micros;
-    totalAmountInMicros = data.total_amount_in_micros;
-    taxBreakdown = data.pricing_phases.base.tax_breakdown;
-    pendingReason = null;
-
-    gatewayParams = {
-      ...gatewayParams,
-      elements_configuration: data.gateway_params.elements_configuration,
-    };
-  }
-
   const handleError = (e: PurchaseFlowError) => {
     const event = createCheckoutFlowErrorEvent({
       errorCode: e.getErrorCode().toString(),
@@ -311,10 +213,8 @@
   {gatewayParams}
   {purchaseOperationHelper}
   {isInElement}
-  {priceBreakdown}
   customerEmail={email ?? null}
   {closeWithError}
   onContinue={handleContinue}
   {onClose}
-  onTaxCustomerDetailsUpdated={refreshTaxCalculation}
 />
